update: main.pyからデータの種類を指定してqd.pyを呼び出すように変更
